LIS-5351: Rewrite all end-to-end integration tests.
The existing test suite was either:

- Relatively mock-heavy, and not testing anything too interesting.
- Relied on locally running nsq infrastructure to complete.

Erik and I took a pass at a complete rewrite of these integration tests
that test the full publisher / subscriber flows, and exercises nsqd failures
and our message delivery semantics. Some other capabilities this PR enables:

1. We can simulate nsqd nodes getting disconnected from the network, and how our client responds.
2. We can now make assertions about message ordering
3. We can stand up and tear down the entire NSQ cluster between every test, for state reset
4. We can now use this to do more aggressive refactorings, including adding some round-robin publishing features we have our eye on.

We are now also requiring Java 8 and above. The client needed to support Java 6 in the past, which
is now EOL.

There should be no meaningful behavior changes in the client, except for hooks that we needed to
expose to make it actually testable.

+++ src/main/java/com/sproutsocial/nsq/Util.java
@@ -58,7 +58,7 @@ class Util {
         }
     }
 
-    public static void sleepQuietly(int millis) {
+    public static void sleepQuietly(long millis) {
         try {
             Thread.sleep(millis);
         }
